requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
telethon==1.31.1
vk-api==11.9.9
aiohttp==3.9.1
python-multipart==0.0.6
cryptography==41.0.5
boto3==1.29.3
schedule==1.2.1